+++ lightly_studio/src/lightly_studio/models/embedding_model.py
@@ -9,17 +9,12 @@ from sqlalchemy import UniqueConstraint
 from sqlmodel import VARCHAR, Column, Field, SQLModel
 
 
-class EmbeddingSpaceDescription(SQLModel):
-    """Description of an embedding space."""
+class EmbeddingModelBase(SQLModel):
+    """Base class for the EmbeddingModel."""
 
     name: str
     embedding_model_hash: str = Field(sa_column=Column(VARCHAR(128), nullable=False))
     embedding_dimension: int
-
-
-class EmbeddingModelBase(EmbeddingSpaceDescription):
-    """Base class for the EmbeddingModel."""
-
     dataset_id: UUID = Field(foreign_key="dataset.dataset_id", index=True)